requirements.txt
pyopengl
glfw
numpy
pillow
matplotlib
tensorflow
opencv-python